Add support for Windows
1. Rearrange the structure of configuration
General setting is moved to autoload/general.vim

2. Remove YouCompleteme
coc.nvim is better then YouCompleteme and I don't work on 32-bit machine
which can't install coc.nvim(no node.js support). So it it time to
remove YouCompleteme.

3. No longer show message of dein.vim in lightline
dein.vim will block vim when installing plugins, so it is not useful to
show message in lightline.