feat(alter): add LSP-based highlight of C/C++ for vim
There are plenty many C++ highlight plugins such as
octol/vim-cpp-enhanced-highlight, bfrg/vim-cpp-modern and so on.
However, all of them are based on regex matching and not accurate
enough.

nvim-treesitter and vim-lsp-cxx-highlight have changed the situation.
nvim-treesitter is based on semantics and vim-lsp-cxx-highlight is based
on LSP. Both of them are robust and smart.

nvim-treesitter only supports
neovim-nightly and vim-lsp-cxx-highlight has some performance problem.
Thus I choose nvim-treesitter for neovim-nightly and
vim-lsp-cxx-highlight for neovim-stable and vim.